move the functions to a different module to make mockable
