+++ .secrets.baseline
@@ -1394,9 +1394,9 @@
         "filename": "src/app/utils/PushNotification.tests.ts",
         "hashed_secret": "91dfd9ddb4198affc5c194cd8ce6d338fde470e2",
         "is_verified": false,
-        "line_number": 67
+        "line_number": 68
       }
     ]
   },
-  "generated_at": "2022-11-30T18:34:01Z"
+  "generated_at": "2022-12-05T20:25:13Z"
 }
